+++ MWS.Data/Entities/UserCategory.cs
@@ -8,7 +8,7 @@ namespace MWS.Data.Entities
         [Key, Column(Order = 1)]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int? id { get; set; }
-        public decimal? user_id { get; set; }
+        public int? user_id { get; set; }
         public int? category_id { get; set; }
     }
 }
